Behave more logically consistent with empty inputs
The mathematical convention is that for any element property, empty sets
satisfy the property (because all elements satisfy the property).  This
is called "vacuous truth".

This is consistent with the behavior of `Iterator::all()` and
`Iterator::any()`.

For example, we will treat:

    is_hiragana("") -> true
    is_japanese("") -> true

Without this change, we would have the behavior that adding characters
to a string could unintuitively change a property from false -> true -> false:

    is_japanese("") -> false
    is_japanese("あ") -> true  // Add あ
    is_japanese("あA") -> false  // Add A

Links:
- https://en.wikipedia.org/wiki/Empty_set
- https://en.wikipedia.org/wiki/Vacuous_truth